Move SideLayout from grid to layout
+ minor improvement
